+++ crates/gem_evm/Cargo.toml
@@ -7,6 +7,7 @@ edition = { workspace = true }
 
 primitives = { path = "../primitives" }
 gem_hash = { path = "../gem_hash" }
+serde_serializers = { path = "../serde_serializers" }
 
 hex = { workspace = true }
 itoa = { workspace = true }
