+++ crates/primitives/src/lib.rs
@@ -92,9 +92,7 @@ pub const DEFAULT_FIAT_CURRENCY: &str = "USD";
 pub mod image_formatter;
 pub use self::image_formatter::ImageFormatter;
 pub mod block_explorer;
-pub mod eip712;
 pub mod encoding_type;
-pub mod erc2612;
 pub mod explorers;
 pub mod solana_token_program;
 pub use self::solana_token_program::SolanaTokenProgramId;
@@ -113,4 +111,3 @@ pub use self::swap_provider::SwapProvider;
 pub mod swap;
 pub mod websocket;
 pub use self::websocket::{WebSocketPriceAction, WebSocketPriceActionType, WebSocketPricePayload};
-pub use eip712::{eip712_domain_types, EIP712Domain, EIP712Field, EIP712Type, EIP712TypedValue};
